Add integration test framework
The integration tests are intended to be executed against a "sacrificial repo" that will have commits, PRs, added to it automatically..

They are used to ensure the correct interaction between spr, git, and
github.

To run these tests first create a sacrificial repo (one time) and add a
file name 'spr.sacrificial' to it. Note this should not be done with any
repo that contains any data you want to keep. Then add it to github and
run the integration tests as follows:
> SACRIFICIAL_REPO=<path to sacrificial repo> go test --tags=integration ./...